src/requirements.txt
openai==1.23.6
langchain==0.2.5
# langchain==0.2.5 requires numpy<2,>=1
langchain-community==0.2.5
ctranslate2==4.3.1
tiledb==0.27.1
tiledb-vector-search==0.2.2
PyMuPDF==1.24.2
InstructorEmbedding==1.0.1
pandas==2.2.2
docx2txt==0.8
python-docx==1.1.0
msg-parser==1.2.0
xlrd==2.0.1
openpyxl==3.1.2
unstructured==0.13.4
psutil==5.9.8
PySide6==6.6.3.1
av==12.0.0
termcolor==2.4.0
pypandoc==1.13
PyYAML==6.0.1
Markdown==3.6
transformers==4.40.2
langchain-huggingface==0.0.3
accelerate==0.29.3
optimum==1.19.1
einops==0.8.0
scipy==1.13.0
timm==0.9.16
sounddevice==0.4.6
platformdirs==4.2.1
rich==13.7.1
openai-whisper
bitsandbytes==0.43.1
webdataset==0.2.86
sentence-transformers==3.0.1

# for ChatTTS
omegaconf==2.3.0
vector-quantize-pytorch==1.14.31

# REQUIRED BY WHISPERSPEECH
encodec==0.1.1
fastcore==1.5.48
fastprogress==1.0.3
HyperPyYAML==1.2.2
# intel-openmp
# mkl
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
soundfile==0.12.1
speechbrain==0.5.16
tbb==2021.13.0
vocos==0.1.0

# for Google TTS
gTTS==2.5.1


